Rename TS_Groups to TS_Large for clarity and consistency
